Move filters to scopes directory
